ivprod/requirements.txt
selenium==4.25.0
schedule==1.2.2
pyautogui==0.9.53
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
flask==3.0.3
requests==2.26.0
numpy==1.21.0  # Optional, include if needed
pandas  # Optional, include if needed
pytest==6.2.4  # Optional, include if needed
tox==3.24.3    # Optional, include if needed
black==21.9b0   # Optional, include if needed
flake8==3.9.2   # Optional, include if needed
selenium==4.25.0
schedule==1.2.2
pyautogui==0.9.53
flask==3.0.3
requests==2.26.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
pytest==6.2.4
tox==3.24.3
black==21.9b0
flake8==3.9.2
isort==5.10.1
flask-socketio==5.3.2
flask-cors==3.0.10 
icalendar==5.0.5 
schedule